backend/pyproject.toml
[project]
name = "youle-backend"
version = "0.1.0"
description = "「有了 / Youle」AI Agent V0 后端 — FastAPI + LangGraph 多智能体反诈短视频 demo"
requires-python = ">=3.12"
dependencies = [
    # === Web 框架 ===
    "fastapi==0.115.6",
    "uvicorn[standard]==0.32.1",
    "python-multipart==0.0.20",
    "python-dotenv==1.0.1",
    "aiofiles==24.1.0",
    "sse-starlette==2.1.3",
    "orjson==3.10.12",

    # === LangGraph 1.x ===
    "langgraph==1.1.9",
    "langgraph-checkpoint-sqlite==3.0.3",

    # === LLM SDK ===
    "anthropic==0.42.0",
    "httpx==0.27.2",

    # === 数据 + 文件 ===
    "pandas==2.2.3",
    "openpyxl==3.1.5",
    "pillow==11.0.0",
    "beautifulsoup4==4.12.3",

    # === 视频 ===
    "ffmpeg-python==0.2.0",

    # === 校验 ===
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",

    # === 可靠性 ===
    "tenacity==8.5.0",

    # === 日志 ===
    "structlog==24.4.0",
    "rich==13.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "pytest-cov==6.0.0",
    "ruff==0.8.4",
    "mypy==1.13.0",
    "respx==0.21.1",
]
v1 = [
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.29.0",
    "alembic==1.13.3",
    "redis==5.0.8",
    "arq==0.26.1",
    "litellm==1.50.4",
    "fastapi-users[sqlalchemy]==14.0.1",
    "slowapi==0.1.9",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM"]
ignore = [
    "E501",  # line-length 由 ruff 的 formatter 处理，不重复报
    "B008",  # FastAPI 习惯用 Depends/File/Form 默认值
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]  # 测试里 assert False 是合法用法

[tool.mypy]
python_version = "3.12"
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
# 第三方库类型不全，先放宽 strict；后续可逐模块收紧
check_untyped_defs = false
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["tests.*", "scripts.*"]
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:websockets.*",
]
